+++ core/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## HEAD
 
+### 🀄 Blockly Workspace
+
+- ⚙️ [v1.1.0] `instanciateEveryBlock`: Add new param `engineData` used to configure the blocks.
+- ⚙️ [v1.1.0] `parseInstructions`: Add the capability to parse new lines.
+
 ### 🤖 PixiJS Engine
 
 - ⚙️ [v1.1.0] `engineGenerator/processors/processorGenerator`: Add the possibility to receive one `CodimoComponent` or an array of `CodimoComponent`, being able to process all of them. Usefull when the activity has more than one actor.
@@ -14,6 +19,10 @@
   - Add an state in which the actor can be, or not, in movement.
   - 🆕 Exportable constant `NO_START_POSITION`.
   - 🆕 Method `setPosition(newPosition: string): void`. It sets the actor in the new position instantly without animation.
+
+### 💅 React UI
+
+- ⚙️ [v1.1.0] `BlocklyApp`: Add `props.blocklyData.doNotUseRootBlock` flag to determine if the Root Blocks is going to be used or not.
 
 ## 1.0.0-alpha.2